# openapi.yaml
openapi: 3.1.0
info:
  title: Silicon IAM API
  version: 1.0.0
  summary: Secure identity, organization governance, application login, and delegated access.
  description: |
    Normative public and privileged HTTP contract for Silicon IAM. PostgreSQL is
    authoritative; webhook delivery never substitutes for online authorization.
    Timestamps are UTC RFC 3339 values. One-time secret responses can be replayed
    for ten minutes only by the same caller using the same route, request digest,
    and Idempotency-Key.
    IAM-managed six-digit OTPs expire after ten minutes. Signup, Carbon login,
    invitation-join, and verified-channel step-up challenges allow ten failed
    verifications, then enter a 60-second cooldown before the current unexpired
    challenge receives a fresh ten-attempt window. Replacement codes inherit
    the partial failed-attempt count or active cooldown.
  contact: { email: auth@teamofsilicons.com }
servers:
  - { url: 'https://backend.iam.teamofsilicons.com', description: Production }
...
paths:
...
  /api/v1/me:
    get:
      tags: [Carbon account]
      operationId: getCurrentCarbon
      security: [{ iamBearer: [] }]
      responses:
        '200':
          description: Current Carbon including self-only contact details.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarbonSelf' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Carbon account]
      operationId: updateCurrentCarbon
      security: [{ iamBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/CarbonProfilePatch' }
      responses:
        '200':
          description: Profile updated.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Idempotency-Replayed: { $ref: '#/components/headers/IdempotencyReplayed' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarbonSelf' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
components:
  securitySchemes:
    iamBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque IAM access token
      description: Carbon, Silicon, or actor-bound application access token; actor class is checked per operation.
    sessionCookie:
      type: apiKey
      in: cookie
      name: iam_session
      description: Secure, HttpOnly, SameSite browser session used only by interactive authorization flows.
    applicationBasic:
      type: http
      scheme: basic
      description: Application ID as username and current versioned application secret as password.
    platformAdminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque platform-admin access token
      description: IAM token whose current Carbon principal has active platform-administrator authority.
    stepUpToken:
      type: apiKey
      in: header
      name: X-Step-Up-Token
      description: Five-minute action-bound token from the step-up verification flow.
    workosSignature:
      type: apiKey
      in: header
      name: WorkOS-Signature
      description: >-
        Comma-delimited `t=<epoch_ms>,v1=<hex_hmac>` value. IAM verifies
        HMAC-SHA256 over `timestamp + '.' + exact raw UTF-8 body` in constant
        time, enforces a 300-second tolerance, and deduplicates provider event ID.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque caller-generated key. Reuse with different bytes returns idempotency_conflict.
      schema: { type: string, minLength: 16, maxLength: 255 }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the expected aggregate version.
      schema: { type: string, pattern: '^"[0-9]+"$' }
...
  headers:
...
    ETag:
      description: Strong aggregate-version ETag.
      schema: { type: string, pattern: '^"[0-9]+"$' }
...
    IdempotencyReplayed:
      description: True when this is a replay of the stored idempotent response.
      schema: { type: boolean }
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, expired, or revoked authentication.
      headers: { WWW-Authenticate: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    Conflict:
      description: Unique, idempotency, replay, terminal-state, or concurrent-state conflict.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    PreconditionFailed:
      description: If-Match does not equal the current aggregate version.
      headers: { ETag: { $ref: '#/components/headers/ETag' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    UnprocessableEntity:
      description: Well-formed request violates field, policy, hierarchy, quorum, or tenant invariants.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    PreconditionRequired:
      description: Required Idempotency-Key, If-Match, or step-up precondition is absent.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, request_id]
          additionalProperties: false
          properties:
            code: { type: string, pattern: '^[a-z][a-z0-9_]{1,63}$' }
            message: { type: string, minLength: 1, maxLength: 500 }
            details: { type: object, additionalProperties: true }
            request_id: { type: string, minLength: 1, maxLength: 128 }
...
    ExistingCarbonId:
      type: string
      pattern: '^[a-z0-9_-]{3,30}$'
      description: Existing Carbon lookup/projection, including immutable legacy IDs containing zero.
...
    TimeZoneId:
      type: string
      minLength: 1
      maxLength: 255
      pattern: '^[A-Za-z0-9._+-]+(/[A-Za-z0-9._+-]+)*$'
      description: Exact identifier that must resolve in the IANA Time Zone Database, such as UTC or Asia/Kolkata.
...
    CarbonPublic:
      type: object
      required: [principal_id, carbon_id, display_name, profile_photo, created_at]
      properties:
        principal_id: { type: string, format: uuid }
        carbon_id: { $ref: '#/components/schemas/ExistingCarbonId' }
        display_name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, 'null'], maxLength: 5000 }
        profile_photo: { type: string, format: uri, maxLength: 2048 }
        created_at: { type: string, format: date-time }
...
    CarbonSelf:
      allOf:
        - { $ref: '#/components/schemas/CarbonPublic' }
        - type: object
          required: [timezone, email, phone_number, status, version, updated_at]
          properties:
            timezone: { $ref: '#/components/schemas/TimeZoneId' }
            email: { type: string, format: email }
            phone_number: { type: string }
            status: { type: string, enum: [active, suspended] }
            version: { type: integer, minimum: 1 }
            updated_at: { type: string, format: date-time }
...
    CarbonProfilePatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        display_name: { type: string, minLength: 1, maxLength: 200 }
        timezone: { $ref: '#/components/schemas/TimeZoneId' }
        description: { type: [string, 'null'], maxLength: 5000 }
        profile_photo: { type: [string, 'null'], format: uri, maxLength: 2048 }
...